+++ src/app/UI/Admin/Setting/Form/FormFactory.php
@@ -77,6 +77,9 @@ readonly class FormFactory
         $form->addText('title_subpage', $this->translator->translate('input_titleSubpage'))
             ->setNullable()
         ;
+        $form->addText('discord_error_log_url', $this->translator->translate('input_discordErrorLogUrl'))
+            ->setNullable()
+        ;
 
         $form->addGroup($this->translator->translate('field_email'));
         $form->addEmail('email', $this->translator->translate('input_email_formReceiver'))